update interface + add CI/CD config
