+++ Project.toml
@@ -1,3 +1,9 @@
+name = "BraketAHS"
+uuid = "aad1d28c-1a3f-47da-806d-2ed76430c1d6"
+authors = ["Yaroslav Kharkov", "Katharine Hyatt"]
+version = "0.0.1"
+
+
 [deps]
 ArgParse = "c7e460c6-2fb9-53a9-8c5b-16f535851c63"
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
@@ -6,9 +12,13 @@ DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
 ITensors = "9136182c-28ba-11e9-034c-db9fb085ebd5"
 JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Braket = "19504a0f-b47d-4348-9127-acc6cc69ef67"
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+Markdown = "d6f4376e-aef5-505a-96c1-9c027394607a"
+InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"
